refactor(core): Extract OAuth functions into oauth.ts (to simplify request-helper-function.ts, preparation for more SSRF improvements)
