+++ packages/core/src/execution-engine/node-execution-context/utils/request-helpers/index.ts
@@ -17,3 +17,4 @@ export {
 	requestWithAuthenticationPaginated,
 	type ResolveValueFn,
 } from './pagination';
+export { refreshOAuth2Token, requestOAuth1, requestOAuth2 } from './oauth';
